Feat : Add field upload image2 and image3 on the form sickleave

+++ app/Controllers/Backend/SickLeave.php
@@ -143,23 +143,38 @@ class SickLeave extends BaseController
         if ($this->request->getMethod(true) === 'POST') {
             $post = $this->request->getVar();
             $file = $this->request->getFile('image');
+            $file2 = $this->request->getFile('image2');
+            $file3 = $this->request->getFile('image3');
 
             $post["submissiontype"] = $this->Pengajuan_Sakit;
             $post["necessary"] = 'SK';
 
             try {
                 $img_name = "";
+                $img2_name = "";
+                $img3_name = "";
+
+                $row = $mEmployee->find($post['md_employee_id']);
+                $lenPos = strpos($row->getValue(), '-');
+                $value = substr_replace($row->getValue(), "", $lenPos);
+                $ymd = date('YmdHis');
 
                 if ($file && $file->isValid()) {
-                    $row = $mEmployee->find($post['md_employee_id']);
-
                     $ext = $file->getClientExtension();
-                    $lenPos = strpos($row->getValue(), '-');
-                    $value = substr_replace($row->getValue(), "", $lenPos);
-                    $ymd = date('Ymd', strtotime($post['submissiondate']));
-
                     $img_name = $this->Pengajuan_Sakit . '_' . $value . '_' . $ymd . '.' . $ext;
                     $post['image'] = $img_name;
+                }
+
+                if ($file2 && $file2->isValid()) {
+                    $ext2 = $file2->getClientExtension();
+                    $img2_name = $this->Pengajuan_Sakit . '_' . $value . '2_' . $ymd . '.' . $ext2;
+                    $post['image2'] = $img2_name;
+                }
+
+                if ($file3 && $file3->isValid()) {
+                    $ext3 = $file3->getClientExtension();
+                    $img3_name = $this->Pengajuan_Sakit . '_' . $value . '3_' . $ymd . '.' . $ext3;
+                    $post['image3'] = $img3_name;
                 }
 
                 if (!$this->validation->run($post, 'sakit')) {
@@ -169,14 +184,32 @@ class SickLeave extends BaseController
 
                     if ($this->isNew()) {
                         uploadFile($file, $path, $img_name);
+
+                        if ($file2 && $file2->isValid())
+                            uploadFile($file2, $path, $img2_name);
+
+                        if ($file3 && $file3->isValid())
+                            uploadFile($file3, $path, $img3_name);
                     } else {
                         $row = $this->model->find($this->getID());
 
-                        if (!empty($row->getImage()) && $post['image'] !== $row->getImage()) {
-                            if (file_exists($path . $row->getImage())) {
+                        if (!empty($post['image']) && !empty($row->getImage()) && $post['image'] !== $row->getImage()) {
+                            if (file_exists($path . $row->getImage()))
                                 unlink($path . $row->getImage());
-                                $file->move($path);
-                            }
+
+                            uploadFile($file, $path, $img_name);
+                        }
+
+                        if (empty($post['image2']) && !empty($row->getImage2()) && file_exists($path . $row->getImage2())) {
+                            unlink($path . $row->getImage2());
+                        } else {
+                            uploadFile($file2, $path, $img2_name);
+                        }
+
+                        if (empty($post['image3']) && !empty($row->getImage3()) && file_exists($path . $row->getImage3())) {
+                            unlink($path . $row->getImage3());
+                        } else {
+                            uploadFile($file3, $path, $img3_name);
                         }
                     }
 
@@ -207,16 +240,36 @@ class SickLeave extends BaseController
                 $list = $this->model->where($this->model->primaryKey, $id)->findAll();
                 $rowEmp = $mEmployee->where($mEmployee->primaryKey, $list[0]->getEmployeeId())->first();
 
-                $path = 'uploads/' . $this->PATH_Pengajuan . '/';
-                $list[0]->setImage($path . $list[0]->getImage());
+                $path = $this->PATH_UPLOAD . $this->PATH_Pengajuan . '/';
+
+                if (file_exists($path . $list[0]->getImage())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage($path . $list[0]->getImage());
+                } else {
+                    $list[0]->setImage(null);
+                }
+
+                if (!empty($list[0]->getImage2()) && file_exists($path . $list[0]->getImage2())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage2($path . $list[0]->getImage2());
+                } else {
+                    $list[0]->setImage2(null);
+                }
+
+                if (!empty($list[0]->getImage3()) && file_exists($path . $list[0]->getImage3())) {
+                    $path = 'uploads/' . $this->PATH_Pengajuan . '/';
+                    $list[0]->setImage3($path . $list[0]->getImage3());
+                } else {
+                    $list[0]->setImage3(null);
+                }
 
                 $list = $this->field->setDataSelect($mEmployee->table, $list, $mEmployee->primaryKey, $rowEmp->getEmployeeId(), $rowEmp->getValue());
-
-                $title = $list[0]->getDocumentNo() . "_" . $rowEmp->getFullName();
 
                 //* Need to set data into date field in form
                 $list[0]->setStartDate(format_dmy($list[0]->startdate, "-"));
                 $list[0]->setEndDate(format_dmy($list[0]->enddate, "-"));
+
+                $title = $list[0]->getDocumentNo() . "_" . $rowEmp->getFullName();
 
                 $fieldHeader = new \App\Entities\Table();
                 $fieldHeader->setTitle($title);
